+++ demos/routes.js
@@ -7,6 +7,7 @@ import SimpleIcons from './cases/SimpleIcons.jsx';
 import Locales from './cases/Locales.jsx';
 import API from './cases/API.jsx';
 import ExtraInfo from './cases/ExtraInfo.jsx';
+import Tooltips from './cases/Tooltips.jsx';
 import BackendData from './cases/BackendData.jsx';
 import DataProvider from './cases/DataProvider.jsx';
 import BackendFilter from './cases/BackendFilter.jsx';
@@ -26,6 +27,7 @@ export const links = [
   ['/locales/:skin', 'Locales', Locales, 'Locales'],
   ['/api/:skin', 'API calls', API, 'API'],
   ['/extra-info/:skin', 'Extra info', ExtraInfo, 'ExtraInfo'],
+  ['/tooltips/:skin', 'Tooltips', Tooltips, 'Tooltips'],
   ['/serverdata/:skin', 'Backend data', BackendData, 'BackendData'],
   ['/data-provider/:skin', 'Saving to backend', DataProvider, 'DataProvider'],
   [